InstAD/requirements.txt
click==8.1.7
faiss-gpu==1.7.2
pillow==8.4.0
timm==1.0.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
scikit-learn==1.5.0
scipy==1.13.1
tqdm==4.66.4
